+++ src/shared/components/case-editor/case-edit-submit/case-edit-submit.component.spec.ts
@@ -12,6 +12,7 @@ import { IsCompoundPipe } from '../../palette/utils/is-compound.pipe';
 import { WizardPage } from '../domain/wizard-page.model';
 import { Wizard } from '../domain/wizard.model';
 import { CaseField } from '../../../domain/definition/case-field.model';
+import { HttpError } from '../../../domain/http/http-error.model';
 import { OrderService } from '../../../services/order/order.service';
 import { aWizardPage } from '../case-edit.spec';
 import { CaseFieldService } from '../../../services/case-fields/case-field.service';
@@ -23,6 +24,7 @@ import { CaseEditPageComponent } from '../case-edit-page/case-edit-page.componen
 import { ProfileService, ProfileNotifier } from '../../../services/profile';
 import { Profile } from '../../../domain';
 import { createAProfile } from '../../../domain/profile/profile.test.fixture';
+import { text } from '../../../test/helpers';
 
 describe('CaseEditSubmitComponent', () => {
 
@@ -93,6 +95,20 @@ describe('CaseEditSubmitComponent', () => {
     },
     params: of({})
   };
+
+  const $SELECT_ERROR_SUMMARY = By.css('.error-summary');
+  const $SELECT_ERROR_HEADING_GENERIC = By.css('.error-summary>h1:first-child');
+  const $SELECT_ERROR_MESSAGE_GENERIC = By.css('.govuk-error-summary__body>p:first-child');
+  const $SELECT_ERROR_HEADING_SPECIFIC = By.css('.error-summary>h3:first-child');
+  const $SELECT_ERROR_MESSAGE_SPECIFIC = By.css('.error-summary>p:nth-child(2)');
+  const $SELECT_CALLBACK_DATA_FIELD_ERROR_LIST = By.css('.error-summary-list');
+  const $SELECT_FIRST_FIELD_ERROR = By.css('li:first-child');
+  const $SELECT_SECOND_FIELD_ERROR = By.css('li:nth-child(2)');
+
+  const ERROR_HEADING_GENERIC = 'Something went wrong';
+  const ERROR_MESSAGE_GENERIC = 'We\'re working to fix the problem. Try again shortly.';
+  const ERROR_HEADING_SPECIFIC = 'The event could not be created'
+  const ERROR_MESSAGE_SPECIFIC = 'There are field validation errors'
 
   describe('Save and Resume disabled', () => {
     pages  = [
@@ -496,6 +512,184 @@ describe('CaseEditSubmitComponent', () => {
     it('should return "Return to case list" text label for cancel button when save and resume enabled', () => {
       let result = comp.getCancelText();
       expect(result).toBe('Return to case list');
+    });
+  });
+
+  describe('Error message display tests', () => {
+    pages = [
+      aWizardPage('page1', 'Page 1', 1),
+    ];
+    const firstPage = pages[0];
+    wizard = new Wizard(pages);
+    const queryParamMapNoProfile = createSpyObj('queryParamMap', ['get']);
+    const snapshotNoProfile = {
+      pathFromRoot: [
+        {},
+        {
+          data: {
+            nonProfileData: {
+              user: {
+                idam: {
+                  id: 'userId',
+                  email: 'string',
+                  forename: 'string',
+                  surname: 'string',
+                  roles: ['caseworker', 'caseworker-test', 'caseworker-probate-solicitor']
+                }
+              },
+              'isSolicitor': () => false,
+            }
+          }
+        }
+      ],
+      queryParamMap: queryParamMapNoProfile,
+    };
+    let PROFILE_OBS: Observable<Profile> = Observable.of(PROFILE);
+    const mockRouteNoProfile = {
+      params: of({id: 123}),
+      snapshot: snapshotNoProfile
+    };
+
+    beforeEach(async(() => {
+      orderService = new OrderService();
+      casesReferencePipe = createSpyObj<CaseReferencePipe>('caseReference', ['transform']);
+      cancelled = createSpyObj('cancelled', ['emit'])
+      caseEditComponent = {
+        'form': FORM_GROUP,
+        'data': '',
+        'eventTrigger': { 'case_fields': [], 'can_save_draft': true },
+        'wizard': wizard,
+        'hasPrevious': () => true,
+        'getPage': () => firstPage,
+        'navigateToPage': () => undefined,
+        'cancel': () => undefined,
+        'cancelled': cancelled,
+      };
+      formErrorService = createSpyObj<FormErrorService>('formErrorService', ['mapFieldErrors']);
+      formValueService = createSpyObj<FormValueService>('formValueService', ['sanitise']);
+
+      profileService = createSpyObj<ProfileService>('profileService', ['get']);
+      profileService.get.and.returnValue(PROFILE_OBS);
+      profileNotifier = new ProfileNotifier();
+      profileNotifier.profile = new BehaviorSubject(createAProfile()).asObservable();
+      profileNotifierSpy = spyOn(profileNotifier, 'announceProfile').and.callThrough();
+
+      TestBed.configureTestingModule({
+        declarations: [
+          CaseEditSubmitComponent,
+          IsCompoundPipe,
+          CaseReferencePipe
+        ],
+        schemas: [NO_ERRORS_SCHEMA],
+        providers: [
+          { provide: CaseEditComponent, useValue: caseEditComponent },
+          { provide: FormValueService, useValue: formValueService },
+          { provide: FormErrorService, useValue: formErrorService },
+          { provide: CaseFieldService, useValue: caseFieldService },
+          { provide: FieldsUtils, useValue: fieldsUtils },
+          { provide: CaseReferencePipe, useValue: casesReferencePipe },
+          { provide: ActivatedRoute, useValue: mockRouteNoProfile },
+          { provide: OrderService, useValue: orderService },
+          { provide: ProfileService, useValue: profileService },
+          { provide: ProfileNotifier, useValue: profileNotifier }
+        ]
+      }).compileComponents();
+    }));
+
+    beforeEach(() => {
+      fixture = TestBed.createComponent(CaseEditSubmitComponent);
+      comp = fixture.componentInstance;
+      de = fixture.debugElement;
+      fixture.detectChanges();
+    });
+
+    it('should display generic error heading and message when form error is set but no callback errors, warnings, or error details', () => {
+      comp.error = {
+        status: 200,
+        callbackErrors: null,
+        callbackWarnings: null,
+        details: null
+      } as HttpError;
+
+      fixture.detectChanges();
+
+      fixture.whenStable().then(() => {
+        const error = de.query($SELECT_ERROR_SUMMARY);
+        expect(error).toBeTruthy();
+
+        const errorHeading = error.query($SELECT_ERROR_HEADING_GENERIC);
+        expect(text(errorHeading)).toBe(ERROR_HEADING_GENERIC);
+
+        const errorMessage = error.query($SELECT_ERROR_MESSAGE_GENERIC);
+        expect(text(errorMessage)).toBe(ERROR_MESSAGE_GENERIC);
+      });
+    });
+
+    it('should display specific error heading and message, and callback data field validation errors (if any)', () => {
+      comp.error = {
+        status: 422,
+        callbackErrors: null,
+        callbackWarnings: null,
+        details: {
+          field_errors: [
+            {
+              message: 'First field error'
+            },
+            {
+              message: 'Second field error'
+            }
+          ]
+        },
+        message: 'There are field validation errors'
+      } as HttpError;
+
+      fixture.detectChanges();
+
+      fixture.whenStable().then(() => {
+        const error = de.query($SELECT_ERROR_SUMMARY);
+        expect(error).toBeTruthy();
+
+        const errorHeading = error.query($SELECT_ERROR_HEADING_SPECIFIC);
+        expect(text(errorHeading)).toBe(ERROR_HEADING_SPECIFIC);
+
+        const errorMessage = error.query($SELECT_ERROR_MESSAGE_SPECIFIC);
+        expect(text(errorMessage)).toBe(ERROR_MESSAGE_SPECIFIC);
+
+        const fieldErrorList = de.query($SELECT_CALLBACK_DATA_FIELD_ERROR_LIST);
+        expect(fieldErrorList).toBeTruthy();
+        const firstFieldError = fieldErrorList.query($SELECT_FIRST_FIELD_ERROR);
+        expect(text(firstFieldError)).toBe('First field error');
+        const secondFieldError = fieldErrorList.query($SELECT_SECOND_FIELD_ERROR);
+        expect(text(secondFieldError)).toBe('Second field error');
+      });
+    });
+
+    it('should not display generic error heading and message when there are specific callback errors', () => {
+      comp.error = {
+        status: 422,
+        callbackErrors: ['First error', 'Second error'],
+        callbackWarnings: null,
+        details: null
+      } as HttpError;
+
+      fixture.detectChanges();
+
+      const error = de.query($SELECT_ERROR_SUMMARY);
+      expect(error).toBeFalsy();
+    });
+
+    it('should not display generic error heading and message when there are specific callback warnings', () => {
+      comp.error = {
+        status: 422,
+        callbackErrors: null,
+        callbackWarnings: ['First warning', 'Second warning'],
+        details: null
+      } as HttpError;
+
+      fixture.detectChanges();
+
+      const error = de.query($SELECT_ERROR_SUMMARY);
+      expect(error).toBeFalsy();
     });
   });
 });